Obj style models in Entity Model.
Added the ability to do obj style models with face lists in entity model,
and wrote a function to generate the model the rendering code uses from
that style of model. Not sure if it's worth the effort. The reason I
initially did it was to be able to do vertex normals in the terrain class,
but that might be easier to just do manually.